+++ crates/ui/widgets/src/dialog/settings_dialog/tab/general_settings.rs
@@ -1,4 +1,5 @@
 use arama_i18n::Locale;
+use arama_env::ThemePreset;
 use arama_env::target_media_type::TargetMediaType;
 
 pub mod message;
@@ -11,6 +12,7 @@ pub struct GeneralSettings {
     sub_dir_depth_limit: u8,
     similarity_threshold: f32,
     locale: Locale,
+    theme: ThemePreset,
 }
 
 impl GeneralSettings {
@@ -19,12 +21,14 @@ impl GeneralSettings {
         sub_dir_depth_limit: u8,
         similarity_threshold: f32,
         locale: Locale,
+        theme: ThemePreset,
     ) -> Self {
         Self {
             target_media_type: target_media_type.to_owned(),
             sub_dir_depth_limit,
             similarity_threshold,
             locale,
+            theme,
         }
     }
 }